# [Java.Interop.Tools.TypeNameMappings] remove Assembly.GetName().Name usage